+++ server/src/controllers/index.js
@@ -1,5 +1,4 @@
 const { clientError, serverError } = require('./error');
-const validateSignUp = require('./validation/validateSignUp');
 const signup = require('./signup');
 
-module.exports = { clientError, serverError, validateSignUp, signup };
+module.exports = { clientError, serverError, signup };
