+++ server/src/controllers/validation/signupSchema.js
@@ -1,0 +1,16 @@
+const joi = require('@hapi/joi');
+// const Boom = require('@hapi/boom');
+
+module.exports = joi.object({
+  name: joi.string().required(),
+  email: joi
+    .string()
+    .email()
+    .required()
+    .pattern(/(\w*@gazaskygeeks.\w*)/),
+  password: joi
+    .string()
+    .regex(/[A-z0-9]{6,}/)
+    .required(),
+  confirmpassword: joi.ref('password'),
+});
