+++ server/src/database/queries/checkEmailExist.js
@@ -2,7 +2,7 @@ const connection = require('../config/connection');
 
 module.exports = (email) => {
   const sql = {
-    text: 'select * from bookinguser where email = $1',
+    text: 'SELECT * FROM BOOKINGUSER WHERE EMAIL = $1',
     values: [email],
   };
   return connection.query(sql);
