+++ server/src/router.js
@@ -1,13 +1,8 @@
 const router = require('express').Router();
-const {
-  clientError,
-  serverError,
-  validateSignUp,
-  signup,
-} = require('./controllers');
+const { clientError, serverError, signup } = require('./controllers');
 const loginValidation = require('./controllers/validation/loginValidation');
 
-router.post('/signUp', validateSignUp, signup);
+router.post('/signUp', signup);
 router.post('/login', loginValidation);
 router.use(clientError);
 router.use(serverError);
